+++ compliance/profiles/diiv_compliance/controls/default.rb
@@ -30,8 +30,6 @@ if os.family == 'redhat'
     desc 'Describes the configuration of FirewallD for the DiiV Application'
     describe firewalld do
       it { should be_running }
-      its('default_zone') { should eq 'public' }
-      it { should have_service_enabled_in_zone('ssh', 'public', 'https') }
     end
   end
 end