+++ metadata.rb
@@ -4,14 +4,13 @@ maintainer_email 'mbutler@optum.com'
 license 'All Rights Reserved'
 description 'Installs/Configures Subsonic'
 
-version '2.0.6'
+version '3.0.0'
 
-chef_version '>= 17'
+chef_version '>= 18'
 supports 'redhat'
 supports 'windows'
 supports 'debian'
-depends 'selinux'
-# depends 'firewall'
+depends 'firewall'
 
 issues_url 'https://github.com/snohio/diiv/issues'
 source_url 'https://github.com/snohio/diiv'